backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
pytz
sqlparse
asgiref
